+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 * Use new `Service<Request>` trait
 
+* Add UDS listening support to `ServerBuilder`
 
 ## [0.2.4] - 2018-11-21